eror di file ini sudah kubenerin

+++ data-mahasiswa.php
@@ -21,45 +21,54 @@
 
 <body>
     <h1>Data Mahasiswa</h1>
-    <?php
-    $mysqli = new mysqli("localhost", "root", "", "fullstack");
-    if ($mysqli->connect_errno)
-        die("Failed to connect to MySQL :" . $mysqli->connect_error);
+  <?php
+$mysqli = new mysqli("localhost", "root", "", "fullstack");
+if ($mysqli->connect_errno) {
+    die("Failed to connect to MySQL: " . $mysqli->connect_error);
+}
 
-    $sql = "select * from mahasiswa";
+$sql = "select * from mahasiswa";
 
-    $stmt = $mysqli->prepare($sql);
-    $stmt->execute();
-    $res = $stmt->get_result();
+$stmt = $mysqli->prepare($sql);
 
-    echo "<table border=1 cell-spacing=0><th>Foto</th> <th>Nama</th> <th>NRP</th> <th colspan='2'>Aksi</th>";
+// Check if the prepare() call failed
+if ($stmt === false) {
+    die("Error preparing statement: " . $mysqli->error);
+}
 
-    while ($row = $res->fetch_assoc()) {
-        echo "<tr>";
-        echo "<td>";
+$stmt->execute();
+$res = $stmt->get_result();
 
-        $fotoMhs = "uploads/" . $row['nrp'] . "." . $row['foto_extension'];
+echo "<table border=1 cell-spacing=0><th>Foto</th> <th>Nama</th> <th>NRP</th> <th colspan='2'>Aksi</th>";
 
-        // cek apakah file benar-benar ada di folder
-        if (file_exists($fotoMhs)) {
-            echo "<img class='foto' src='" . $fotoMhs . "' alt='poster'>";
-        } else {
-            echo "<span class='teks-merah'>Poster tidak ditemukan</span>";
-        }
+while ($row = $res->fetch_assoc()) {
+    echo "<tr>";
+    echo "<td>";
 
-        echo "</td>";
+    $fotoMhs = "uploads/" . $row['nrp'] . "." . $row['foto_extension'];
 
-        echo "<td>" . $row['nama'] . "</td>";
-        echo "<td>" . $row['nrp'] . "</td>";
-
-        echo "<td><a href='edit-mahasiswa.php?npk=" . $row['nrp'] . "'>Edit</a></td>";
-        echo "<td><a href='delete-mahasiswa.php?npk=" . $row['nrp'] . "' onclick='return confirm(\"Yakin ingin menghapus mahasiswa ini?\");'>Delete</a></td>";
-        echo "<input type='hidden' name='nrp_lama' value='" . $row['nrp'] . "'>";
-        echo "</tr>";
+    if (file_exists($fotoMhs)) {
+        echo "<img class='foto' src='" . $fotoMhs . "' alt='poster'>";
+    } else {
+        echo "<span class='teks-merah'>Poster tidak ditemukan</span>";
     }
 
-    echo "</table>";
-    ?>
+    echo "</td>";
+
+    echo "<td>" . $row['nama'] . "</td>";
+    echo "<td>" . $row['nrp'] . "</td>";
+    echo "<td><a href='edit-mahasiswa.php?npk=" . $row['nrp'] . "'>Edit</a></td>";
+    echo "<td><a href='delete-mahasiswa.php?npk=" . $row['nrp'] . "' onclick='return confirm(\"Yakin ingin menghapus mahasiswa ini?\");'>Delete</a></td>";
+    echo "<input type='hidden' name='nrp_lama' value='" . $row['nrp'] . "'>";
+    echo "</tr>";
+}
+
+echo "</table>";
+
+// Close the statement and connection
+$stmt->close();
+$mysqli->close();
+?>
 </body>
 
 </html>
